.bashrc: Prefixable PS1
I liked the old one-line PS1 better than the one I came up with
earlier today, but it was hard to prefix it with information, such as
a Git branch or a virtualenv name. This one is still a bit difficult,
and I should come up with a way to make it more flexible, but it's a
good start for now.