bash: Minimal bash prompt
I'm kind of getting tired of the "fancy" PS1 I've been using the past
few months (as well as how it looks with the Solarized color scheme I'm
using in my terminal, and sometimes Solarized in general, since the
emphasized colors are mostly just grey, but I digress), so I'm going to
try a more minimal prompt. To start off, it's just a '$'. I might add
more bit by bit; for example, right now, when I'm in a Python
virtualenv, the only indication is that the prompt turns color (to
yellow) -- it might be nice to show the name, too.

If I decide to add more information to the prompt, I'll probably break
it into two lines, and maybe even write a function to build a prompt
for me (the benefit being that I can also hide the color variables I
use inside that).

Anyway, ideas, right? First step is just using a minimal PS1.