Revert "bash: Use `tput` to set terminal colors"
Unfortunately, this interferes with the terminal display when, e.g.,
cycling through the shell history. I _think_ this is due to the fact
that bash is unable to correctly calculate the width of the prompt, as
described in [1]. I'm going to revert these changes until I can figure
out how to do it properly.

This reverts commit 750c73b9f3adef3f2cb1baa33f37679f16422f08.

  [1] http://unix.stackexchange.com/a/90876/57970