bash: Ensure that cowsay output never wraps
`cowsay -n` can cause weird drawing issues if the text piped to it is
longer than the terminal width. Luckily, we can use `cut -c` to ensure
that the text is wrapped to the maximum width of the terminal, in the
case of long lines.

Thanks to [1] for the helpful tip.

  [1] http://superuser.com/a/58812/11374