bash: Remove ~/.rubies/Current symlink
Having a "Current" symlink, as I do with ~/.pythons and ~/.scalas,
interferes (a bit) with chruby's listing of installed Ruby versions --
it lists "Current" as an option. I like having that symlink to quickly
see what version of Ruby is the default, but it should be fine to get
rid of it and just use the latest version of Ruby that I have installed
as the default.