bash: Remove TODO item
I did some research. In a nutshell, bash sources ~/.bash_profile for
login shells, and ~/.bashrc for non-login shells (but only if they're
not interactive). So basically, if you log in via a console, you get
~/.bash_profile, and if you just start a new terminal window, you get
~/.bashrc. Except on OS X, where every new terminal in Terminal.app is a
login terminal.

But, of course, I rarely log in to a machine via a console (normally I
use a GUI). So really, it's best to put everything in ~/.bashrc and just
source that from ~/.bash_profile.