bash: Create local variables without a function
I'd like to avoid "polluting" the shell process with useless variables
that I create and use only in ~/.bashrc. Unfortunately, the `local`
keyword can only be used to control the visibility of variables in
functions. However, I believe you can wrap part of a bash script in {
and } to essentially create an anonymous function, within which you
_can_ use the local keyword to control visibility.

I'm at work right now (where I don't currently use bash), but I'm going
to try this at home and see if it works. If it does, great! If not, I
_think_ I can also run this part of the script in a subshell to prevent
variables from being visible outside of the script...