bash: Configure ~/.bashrc to work with Linux/ILM
I'd like to start using bash on my Linux workstation at ILM (currently I
use tcsh there, as that's the default), but ~/.bashrc was set up in a
very Mac-centric way. These changes should make it more cross-platform,
though. Originally I'd planned to separate the Mac-only stuff from the
Linux-only stuff (and the ILM-specific stuff) and put them into seprate
files, such as ~/.bash_env.linux and ~/.bash_env.darwin, that were then
conditionally sourced in ~/.bashrc. However, I think so much of the
stuff in ~/.bashrc is general (or "global") that I can make do with a
few conditional config statements in ~/.bashrc itself, despite the fact
that it adds a bit of complexity to what was once a fairly simple
~/.bashrc file. If it gets more complicated I may consider splitting it
out into separate files, but this should suffice for now.