bash: Recreate `readlink -f` on OS X
OS X (and Darwin, and, I believe, BSDs in general) do not support the
`-f` flag to `readlink` like GNU's version does on Linux. `abspath` was
a hacked attempt to essentially recreate that behavior on OS X. However,
it only worked with directories. This function, suggested by [1], should
basically defer to `/usr/bin/readlink` _unless_ the `-f` flag is passed,
in which case it will do the same thing as `readlink -f` on GNU/Linux.

  [1] http://stackoverflow.com/a/1116890/28804