Revert "bash: Save history after every command"
Unfortunately, this interfered with Terminal.app on Mac OS X in some
way. I have Terminal.app setup to start new tabs in the working
directory of the previous tab. Unfortunately, setting `$PROMPT_COMMAND`
in this way made tabs start in my home directory instead. Not sure why,
but removing this command fixes the issue. Sad, because this seemed like
a good idea, but oh well.

This reverts commit 97c69595644b56b2c768f1fe31982d313cd56d21.