bash: Try out a fancier prompt
This one has username, hostname, and PWD in the prompt. A little bit
fancier than before. Not sure I'll dig it at home, but maybe I can make
it work with a few modifications. Eventually I might add information
like Python virtualenv and Git branch to the prompt as well... On the
other hand, I _do_ like a good minimalist prompt, so who knows.

(I should also maybe develop a set of functions for adding or removing
"components" to the bash prompt, but I'll save that for later.)

I was inspired by a post on Steve Losh's website [1].

  [1] http://stevelosh.com/blog/2009/03/mercurial-bash-prompts/