bash: Modify `vv` so virtual environment name works with new $PS1
The new $PS1 was introduced in 414d502.

The prompt is getting fuck-long; might have to remove the username or
something, especially if I want to add additional things into the
prompt, such as the current Git branch. But I'll cross that bridge when
I come to it.

Or maybe I'll just go back to my nice old minimalist prompt.