bash: Escape \$ in $PS1
I actually want to use the \$ variable that works in bash's $PS1 -- it
will swap out a # instead of a $ when I am logged in as root. This isn't
terribly important to have, since I don't use this particular $PS1 with
root, but hey, it's the "right thing" to do so I might as well do it.
However, it means that I have to escape the \; before, a literal "$" was
being written into $PS1, since I was just using \$.