bash: Make directory in PS1 blue
I use a bunch of commands, particular git subcommands, that output a lot
of green to the screen; it's easier to pick out the prompt when it's
blue. Plus, directories when listed with `ls` are blue (at least, on my
OS X machine they are), so it makes sense for the directories in the
prompt to be blue, too. Might consider changing this, though. (Also...I
did kind of like the green better, but whatever. Hey, maybe purple!)

(Might also add things like an indicator of the status of the last
command as well, so this could be a-changing anyway.)