bash: Remove `brew` function
I wrote a custom script for my ILM workstation that sort of mimics the
behavior of `brew` on OS X; as a result, I don't really need this custom
function anymore, as I was only using it on my ILM workstation anyway.

This function or the custom script I wrote may be useful in the future
if I move to other platforms more frequently, in which case I may copy
my custom "poor man's `brew`" (aka `pißwasser`) into this repo. We'll
see.