bash: pißwasser: A poor man's Homebrew
`pißwasser` is designed to provide a barebones simulacrum of `brew` on
systems without Homebrew installed (e.g., Linux) so that `brew` can
still be used in bash scripts with impunity. It is _not_ a complete
recreation of OS X's Homebrew, nor is it intended to be so.

`pißwasser` was created so I could still use `brew` in my bash config
files, even on systems without Homebrew.