bash: Improve readability of `greet`
I may just end up moving this to the end of ~/.bashrc, as that will work
on both OS X and Linux (even though it's "more" correct to put it in
~/.bash_profile) with this sort of construct:

    if is_login_shell; then  # remember how do detect login in bash
      if hash lolcat 2>/dev/null && if hash cowthink 2>/dev/null; then
        # do greeting
      fi
    fi

I'm also not sure I want the function to return 1 if it fails; that
makes sense in the context of a function, but not necessary for _my_
purposes, which is just to display a greeting when launching bash -- I
may want it to just exit silently if the proper programs can not be
installed (I don't know how exiting 1 really affects bash's startup
procedure). But I can work that out later, I suppose.