git: Fix `git alias` display
`git alias` basically works like this: It gets all alias.* config
variables, removes the "alias." part, replaces the first space with an
underscore (_), then uses the underscore as a delimter for `column` to
neatly print the output. Great, huh? There was one problem, though: A
git alias may contain an underscore, which caused problems with
`column`! So I decided to change the delimeter to `+` instead. That's
great...except the definition of `git alias` itself contains a "+",
which caused `column` to do weird things on that entry.

`xxd` to the rescue! `echo $ascii_value | xxd -r -p` will print out the
character for a given ASCII code point, which allows me to create the
`git alias` definition without actually having a plus sign (+) in the
definition. Cool!