python: Ignore ANSI codes when calculating prompt length
Readline includes the ANSI codes when calculating the length of the
prompt, which means that the display when cycling through
previously-entered lines is displayed incorrectly (it doesn't completely
delete the line in some cases). More info is available at [1], but
suffice it to say that surrounding the ANSI codes in the Readline
ignore/unignore control characters (\001 and \002) solves the problem.

  [1] https://stackoverflow.com/a/9468954/28804