bash: Set $PIP_CONFIG_FILE on macOS
By default on macOS, Pip looks in ~/Library/Application
Support/pip/pip.conf for its config file, which is a really bizarre
place to look for it, but is aligned with the bizarre pedantry commont
to the Python world. Luckily, this nutty decision can be changed by
using the $PIP_CONFIG_FILE environment variable, which we will set to
~/.config/pip/pip.conf, to avoid having to use a different symlink on a
platform-by-platform basis.