bash: Allow completions on a per-host basis
I don't have a version of Haskell compiled for CentOS 6, so the
compleat_setup completion cannot run on CentOS 6 machines.

Separating configs by host is getting to be a _bit_ cumbersome and I'm
not sure I like it more than my previous solution of setting everything
dynamically, but I've already gone down the rabbit hole so I'll keep
working with it for now.