python: Modify startup script to work with Django
When a Python shell is opened using `python manage.py shell` (in
Django), the Python startup file is executed directly, and thus the
`__file__` variable is not available, which obviously caused problems in
this script. If `__file__` is not defined, fall back to assuming the
startup script is at `$XDG_CONFIG_HOME/python/statup.py` and try to
create a history file relative to that file.

Obviously this will cause problems should the startup.py file be located
anywhere other than `$XDG_CONFIG_HOME/python/startup.py`, but I'll cross
that bridge when I come to it.