bash: Convert `greet` function to native program
My shell takes a long time to start up now. Perhaps using a native C
program to print my greeting will improve the startup time.

(Note that in practice, a native binary does not seem to help a whole
lot, but maybe I can fiddle with the shell a bit to improve the startup
time anyway.)