bash: Revert static definition of $PATH
It was getting a bit unwieldy to manage and harder to integrate with
external and third-party scripts that are able to manage the $PATH
variable using `eval`. I'm going to try doing it the old-fashioned way
again.

This removes some Linux-specific customizations but I can hack those in
again later if need be.