bash: Update `bi` alias to modern standards
This is the preferred mechanism for setting the bundle installation path
on the latest versions of Bundler.

I think I could also set the $BUNDLE_PATH environment variable to
`.bundle` so this would be configured globally, or set it in my global
Bundler config, but, meh, I rarely use Ruby anymore anyway so who cares?

See [1] for how to set this globally.

  [1] https://bundler.io/v2.4/man/bundle-config.1.html