Open new tabs in current working directory in macOS Terminal
Terminal.app in macOS has an option to open new tabs in the same working
directory as the current tab, but that has not been working for some
time. Thanks to [1], I learned that macOS accomplishes this by calling a
function called `update_terminal_cwd` via `$PROMPT_COMMAND` [1], but
since I overrode `$PROMPT_COMMAND` in d5dfdda, this feature was no
longer working. I now call the default `$PROMPT_COMMAND` function in
addition to my own to ensure that it works properly on macOS.

Truth be told, I'm not sure why `update_terminal_cwd` needs to be called
for this feature to work, as it appears that `update_terminal_cwd`
simply changes the title of the tab to include the working directory,
but I do know that it won't work if that function is not called.

I also don't really like how it prints the full path of the working
directory in the tab title, so I might end up reverse-engineering
`update_terminal_cwd` and write my own, if possible, but it might hook
into Apple Terminal in a way that is not obvious to me nor possible to
override.

  [1] https://apple.stackexchange.com/a/128999/411