git: Configure Kaleidoscope as default diff and merge tool
I just installed this tool (the old 2.4.6 version because the company
ridiculously wants me to pay a _monthly subscription_ for the newest
version, lol) and when I automatically configured it, this was the
configuration I got by default...so I assume that it works fine.

To remove it as the default tool, remove everything but the
difftool.Kaleidoscope and mergetool.Kaleidoscope sections.