haskell: Don't warn about type defaults in GHCI
GHCI was compiled with -Wtype-defaults, and so any time a result
returned a Num, it would print a warning that the type variable was
defaulting to Integer. For example, simply doing

    map (* 2) [1..10]

would print such a warning, since the resulting type is actually [Num],
not [Integer], so GHCI has to infer that it is [Integer] in order to
show the result. Annoying, so let's disable that. (I'm sure this will
come back to bite me in the ass at some point.)