python: Never store history next to startup.py file
The logic here was a bit weird and was probably introduced over time as
I iterated over this file. Essentially I only got a NameError because I
was trying to use the `py_config_home` variable before it had been
initialized. I think I intended to pass `startup_py` to
`os.path.dirname()`, but truth be told, I always want to use
`$XDG_STATE_HOME` and defer to `~/.local/state` as a backup, and never
store the history file in `~/.config/python`, so I might as well just
remove that bit instead of trying to fix it somehow.