bash: Pipe httpstatus through less
I rarely if ever want to keep the page open in Lynx, so just dump it to
the screen and page through it using less.

I am also now retrieving the page as utf-8 because it has special "××"
characters that don't display properly unless the `-display_charset`
flag is specified.

I'm also removing the first couple of lines because they're an annoying
nav banner that I don't really need to see.