bash: Replace screen-size script with binary
This binary is written in Swift and uses NSScreen to get screen
resolution information instead of using AppleScript. The AppleScript
solution had a few limitations: it could only get the resolution of the
main screen, and it did not always return the correct resolution on
modern "Retina" displays. This binary should be able to get the
resolution of the screen of whatever window is currently focused,
although it has not yet been tested on systems with multiple screens.

I hate to include binaries in the repository, but this one is relatively
small. I'm not sure how portable it will be, especially to older
systems, and it might not be portable to Intel Macs at all (yes, I still
have an Intel Mac) -- and of course, it won't work on Linux at all -- so
I might need a more flexible solution in the future, but I'll cross that
bridge when I come to it.

I also hate including Mac binaries on Linux, but hey, cross that bridge
when I come to it, too.

This binary is also kind of big for what it does, and links to a lot of
libraries. If portability proves to be an issue, I might write it in
good old Objective-C instead of Swift.