bash: Revert to AppleScript method for getting screen resolution
I was unable to properly get the screen size on systems with multiple
displays; this old AppleScript method only shows the resolution of the
primary display (which I think is the screen with the Dock, but I'm not
sure; it might be the screen defined as the primary in System Settings).

Supposedly `+[NSScreen mainScreen]` will get the resolution of the "main
screen", which is the screen with the active window, but when I rewrote
the program in Swift, and then Objective-C, to use `+[NSScreen
mainScreen]`, I could still only get the resolution of the primary
display. I tried using `CGMainDisplayID()` from CoreGraphics to get the
main display, but I still only got the resolution of the primary
display.

Then it hit me: since I am launching these programs from the Terminal,
and they are command-line programs, they don't have a window, so they
have no way to detect which screen they are running on, so I imagine all
the techniques I was using fell back to returning the resolution of the
primary display.

What I need to do to solve this problem from the command line is to
somehow detect which screen the controlling Terminal window is on, which
means I need to be able to get the windows of the Terminal from which
this program was launched. Something like this:

  1. Launch `screen-size`.
  2. Within `screen-size`, figure out which Terminal window it was
     launched from.
  3. #2 is hard, because its parent process is not Terminal.app, but
     rather bash, and on macOS, I believe bash's parent process is
     login...but the parent process of that is Terminal.app, so I can
     maybe follow that up the chain to get Terminal.app's PID, if need
     be.
  4. Somehow get Terminal.app's windows, which may or may not be
     possible.
  5. Assuming #4 works, somehow figure out which screen the window is
     on. Which might be hard, because a window can be on multiple
     screens, so I might have to do something like:
       1. Get the frames in the global coordinate space of all screens.
       2. Find out which screen frames the window's frame intersects.
       3. Pick one if there are multiple? And get its resolution -- this
          is the easy part!

I'm not even sure if it's possible for a command-line program to even
get the windows of an application, and I'm not sure if it's feasible for
a window to determine which screen(s) it is on.

Regardless, the point is moot because I'm not even using a multi-screen
system anymore...so this old AppleScript will continue to work fine. But
it was cool to learn about macOS's display primitives anyway! Learning
on our own is fun and the most rebellious thing we can do in the face of
ubiquitous AI.