bash: Alias 'binld' to 'blib'
Sometimes I write some great functions and completely forget about them.
Turns out I had already defined `blib` long ago, as `binld`, and I
probably haven't used it since because I completely forgot about it. I
was tempted to alias `blib` to `binld`, but `blib` is _a little_ more
robust, as it checks to see if the binary exists before trying to list
its linked libraries, which yields a nicer error message in the case in
which the binary does not exist.