ripgrep: Format output when piping to less
When piping to less (i.e., not printing to a tty), I'd like the same
output that I see when sending output to stdout -- just like ack.

I suppose this could get annoying if I actually want to work with the
output in another tool, but I can disable this config by passing
`--no-config` when invoking `rg`.