bin: New script: `timescale`
In many of Charles Stross's science fiction novels, durations are
expressed exclusively in seconds: second, kilosecond, megasecond, and so
forth. Unfortunately it is hard for my puny human mind to grok these
seconds compare to units I am more used to, such as minutes, hours,
days, and centuries. This is a simple script to convert durations
expressed exclusively in seconds to more human-friendly units.

This is somewhat of a proof-of-concept script to show how data types
might be used to express these durations, too.