Remove `rgp` script in favor of `rg` function
I read somewhere that `rgp` is the script the creator of ripgrep uses to
automatically invoke paging, but it's cumbersome to remember to type
"rgp" in place of "rg"; plus, I'm a tabletop roleplaying aficionado, so
I often type "rpg" by mistake. However, I cannot rename this script to
`rg` and then invoke the OG `rg` using `command` because the script will
be picked up by `command` instead, so using a function to essentially
alias `rg` to `rg ... | less` is the best thing I can do.